backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
httpx==0.28.1
pandas==2.3.2
nubra-talib==0.1.4
protobuf==6.33.0
